[fixed] Fix for bug507.
In the case when `<Input />` type is `submit`,
bootstrap style set in `bsStyle` property
goes down to inner `<Button />` conponent.

Consequently default bootstrap styles for `<Button />`
are not suited for `<FormGroup />` component that
is the root element for `<Input />` component.

Therefore in that case we just need to prevent `bsStyle` passing on
to <FormGroup />.

Tests added and temporarily suspended test is unskipped.