[fixed] Remove cross import between Button & ButtonInput
They were importing each other which causes an issue when using
ButtonInput for server side react rendering under node. Possibly not an
issue when they are bundled into the same client side file.

We remove the import Button and move the definition of the 'types' into
Button which is then referenced from ButtonInput as that already has to
import Button.

We keep the ButtonInput.types to preserve the interface.